Flatten node cursors
