+++ community/kernel/src/main/java/org/neo4j/kernel/impl/api/store/StoreStatement.java
@@ -78,7 +78,7 @@ public class StoreStatement implements StorageStatement
             @Override
             protected StoreSingleNodeCursor create()
             {
-                return new StoreSingleNodeCursor( nodeStore.newRecord(), neoStores, StoreStatement.this, this,
+                return new StoreSingleNodeCursor( nodeStore.newRecord(), neoStores, this,
                         recordCursors, lockService );
             }
         };
